requirements.txt
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
Deprecated==1.2.13
idna==3.4
pycparser==2.21
PyGithub==1.57
PyJWT==2.6.0
PyNaCl==1.5.0
PyYAML==6.0
requests==2.28.1
urllib3==1.26.13
wrapt==1.14.1
